tools/archsync/pyproject.toml
[project]
name = "archsync"
version = "0.1.0"
description = "Interface-first architecture diagrams and diffs generated from code with local LLM enrichment"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "ArchSync Contributors" }]
dependencies = [
  "typer>=0.12.3",
  "pyyaml>=6.0.1",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "watchdog>=4.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "playwright>=1.48.0",
  "ruff>=0.7.0",
]

[project.scripts]
archsync = "archsync.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "e2e: browser based end-to-end tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
